requirements.txt
Django==1.11
djangorestframework==3.6.2
psycopg2==2.8.6
pytz==2021.1
